Remove unneeded insert(statement:) method
